requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.41.1
torch==2.3.0
pytesseract==0.3.10
pdfplumber==0.10.2
scipy==1.13.0
scikit-learn==1.5.0
pandas==2.2.2
spacy==3.7.4
scispacy==0.5.4
sentencepiece==0.2.0
Pillow==10.3.0
opencv-python==4.9.0.80
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.1
